Add an option for minimum interval between notifications
This is useful if you don't want certain apps to constantly send notifications
to your Pebble, when you are already aware of them. This is similar to the
minimum interval between notifications, except it won't even send the
notification to the Pebble at all. It's just a quick patch though.

I'm not sure if this is a feature that you are interested in, but it's something
which I needed, and figured I might as well offer. It's helpful when you want
to try to help save Pebble battery, while having many IM conversations.

Possible room for improvement would be "notify once for this app" feature,
so that if an app has the "top" notification, it wouldn't re-notify the Pebble.
That might already be part of the app though.